pyproject.toml
[project]
name = "js-flask-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
]
